PR-CLEANUP-1 — over-split helper modules + 1-release-grace aliases
Three module files + one core.paths alias removed without behaviour
change:

- core/agent/loop/loop.py (30 LOC re-export shim) — 0 production callers,
  package __init__ already re-exports the same symbols. The one test
  caller (tests/plugins/petri_audit/test_skeleton.py) was migrated to
  core.agent.loop.agent_loop.
- core/agent/loop/_announce.py (41 LOC) + core/agent/loop/_decomposition.py
  (81 LOC) absorbed into core/agent/loop/_helpers.py. Both were <100 LOC
  with exactly one caller (agent_loop.py); the sibling-module split was
  over-engineered for the file size and obscured locality.
  _response.py (184 LOC, distinct streaming responsibility) stays separate.
- core.paths.GLOBAL_JOURNAL_DIR legacy alias removed (1-release grace
  expired; 0 external callers).

Docstring/comment drift cleaned in core/agent/plan.py, agent_loop.py,
and tests/test_model_split.py so references point at the new
_helpers host instead of the deleted modules.

Codex MCP verified (read-only): zero live references to the deleted
modules + alias; absorbed function bodies match the pre-deletion
semantics exactly.
